Refactor UdapClient.  Reduce duplicated code.
