+++ _tests/UdapServer.Tests/Common/TestExtensions.cs
@@ -28,12 +28,11 @@ public static class TestExtensions
         builder.Services.AddScoped<IUdapClient>(sp =>
             new UdapClient(
                 pipeline.BrowserClient,
-                sp.GetRequiredService<TrustChainValidator>(),
+                sp.GetRequiredService<UdapClientDiscoveryValidator>(),
                 sp.GetRequiredService<IOptionsMonitor<UdapClientOptions>>(),
-                sp.GetRequiredService<ILogger<UdapClient>>(),
-                sp.GetRequiredService<ITrustAnchorStore>()));
-        
-            
+                sp.GetRequiredService<ILogger<UdapClient>>()));
+
+        builder.Services.TryAddSingleton<UdapClientDiscoveryValidator>();
         builder.Services.TryAddSingleton<UdapClientMessageHandler>();
         builder.Services.TryAddSingleton<IPostConfigureOptions<TieredOAuthAuthenticationOptions>, TieredOAuthPostConfigureOptions>();
         return builder.AddOAuth<TieredOAuthAuthenticationOptions, TieredOAuthAuthenticationHandler>(
