+++ examples/FhirLabsApi/appsettings.Development.json
@@ -81,8 +81,8 @@
         "SignedMetadataConfig": {
           "RegistrationSigningAlgorithms": [ "ES384" ],
           "TokenSigningAlgorithms": [ "ES384" ],
-          "Issuer": "http://localhost/fhir/r4",
-          "Subject": "http://localhost/fhir/r4",
+          "Issuer": "https://localhost:7016/fhir/r4",
+          "Subject": "https://localhost:7016/fhir/r4",
           "AuthorizationEndPoint": "https://host.docker.internal:5002/connect/authorize",
           "TokenEndpoint": "https://host.docker.internal:5002/connect/token",
           "RegistrationEndpoint": "https://host.docker.internal:5002/connect/register"
